Fix bug in PackFileBackend
As things were implemented, we weren’t careful to ensure that the pack
file had safely landed before writing the index file. In turn, if a
backup was interrupted or failed, we could have ended up with an index
file that claimed we had various hashes saved even though this wasn’t
true.